% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseNestingArea.R
\docType{data}
\encoding{UTF-8}
\name{DatabaseNestingArea}
\alias{DatabaseNestingArea}
\title{Database of RMU for marine turtles}
\format{
A dataframe with raw data.
}
\usage{
DatabaseNestingArea
}
\description{
Database of RMU for marine turtles - Version 2010
}
\details{
Database of RMU for marine turtles
}
\examples{
\dontrun{
library(embryogrowth)
data(DatabaseNestingArea)
}
}
\references{
\insertRef{7332}{embryogrowth}\cr
}
\author{
Marc Girondot \email{marc.girondot@universite-paris-saclay.fr}

Maria Sousa Martins \email{maria.esmartins@gmail.com}
}
\keyword{datasets}
