% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseTSD.version.R
\name{DatabaseTSD.version}
\alias{DatabaseTSD.version}
\title{Version of database of TSD information for reptiles}
\usage{
DatabaseTSD.version()
}
\value{
The date of the lastest updated version
}
\description{
Return the date of the most recent update of the database.
}
\details{
Database of TSD information for marine turtles
}
\examples{
\dontrun{
library(embryogrowth)
DatabaseTSD.version()
}
}
\seealso{
Other Functions for temperature-dependent sex determination: 
\code{\link{DatabaseTSD}},
\code{\link{P_TRT}()},
\code{\link{TSP.list}},
\code{\link{plot.tsd}()},
\code{\link{predict.tsd}()},
\code{\link{stages}},
\code{\link{tsd_MHmcmc_p}()},
\code{\link{tsd_MHmcmc}()},
\code{\link{tsd}()}
}
\author{
Marc Girondot \email{marc.girondot@u-psud.fr}
}
\concept{Functions for temperature-dependent sex determination}
