% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hist.NestsResult}
\alias{hist.NestsResult}
\title{Show the histogram of temperatures with set of nests}
\usage{
\method{hist}{NestsResult}(x, series = "all", ...)
}
\arguments{
\item{x}{Results obtained after searchR}

\item{series}{Series to be used, logical (TRUE ou FALSE), numbers or names. If "all", all series are used.}

\item{...}{Parameters used by hist function (example main="Title")}
}
\value{
A list with an histogram object with information on histogram or
NULL if no series was selected and the complete set of temperatures used.
}
\description{
Show the histogram of temperatures with set of nests
hist(data)
}
\details{
hist.NestsResult shows the histogram of temperatures with set of nests
}
\examples{
\dontrun{
library(embryogrowth)
data(resultNest_4p)
h <- hist(resultNest_4p, series=c(1:5))
}
}
\author{
Marc Girondot \email{marc.girondot@u-psud.fr}
}

