% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.NestsResult.R
\name{logLik.NestsResult}
\alias{logLik.NestsResult}
\title{Return Log Likelihood of a fit generated by searchR}
\usage{
\method{logLik}{NestsResult}(object, ...)
}
\arguments{
\item{object}{A result file generated by searchR}

\item{...}{Not used}
}
\value{
The Log Likelihood value of the fitted model and data
}
\description{
Return Log Likelihood of a fit generated by searchR
}
\details{
logLik.NestsResult Return Log Likelihood of a fit
}
\examples{
\dontrun{
library(embryogrowth)
data(resultNest_4p_SSM)
logLik(resultNest_4p_SSM)
AIC(resultNest_4p_SSM)
}
}
\author{
Marc Girondot
}
