% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.STRN.R
\name{logLik.STRN}
\alias{logLik.STRN}
\title{Return Log Likelihood of a fit generated by STRN}
\usage{
\method{logLik}{STRN}(object, ...)
}
\arguments{
\item{object}{A result file generated by STRN}

\item{...}{Not used}
}
\value{
The Log Likelihood value of the fitted model and data
}
\description{
Return Log Likelihood of a fit generated by STRN
}
\details{
logLik.STRN Return Log Likelihood of a fit
}
\examples{
\dontrun{
library(embryogrowth)
data(resultNest_4p_SSM)
logLik(resultNest_4p_SSM)
AIC(resultNest_4p_SSM)
}
}
\author{
Marc Girondot
}
