% Generated by roxygen2 (4.0.2): do not edit by hand
\name{merge.mcmcComposite}
\alias{merge.mcmcComposite}
\title{Merge two mcmcComposite results}
\usage{
\method{merge}{mcmcComposite}(x, y, ...)
}
\arguments{
\item{x}{A mcmcComposite result}

\item{y}{A mcmcComposite result}

\item{...}{Other parameters for merge function (not used)}
}
\value{
A mcmcComposite result
}
\description{
Merge two mcmcComposite results
}
\details{
merge.mcmcComposite Merge two mcmcComposite results
}
\examples{
\dontrun{
data(result_mcmc_4p)
# Of course, normally you should do it with two different mcmc objects !
result_mcmc <- merge(result_mcmc_4p, result_mcmc_4p)
}
}
\author{
Marc Girondot
}

