% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\docType{data}
\name{nest}
\alias{nest}
\title{Timeseries of temperatures for nests}
\format{A dataframe with raw data}
\usage{
nest
}
\description{
Timeseries of temperatures for nests
}
\details{
Timeseries of temperatures for nests
}
\examples{
\dontrun{
library(embryogrowth)
data(nest)
}
}
\author{
Marc Girondot \email{marc.girondot@u-psud.fr}
}
\references{
Girondot, M. & Kaska, Y. 2014. A model to predict the thermal 
         reaction norm for the embryo growth rate from field data. Journal of
         Thermal Biology. 45, 96-102.
}
\keyword{datasets}

