% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotR_hist.R
\name{plotR_hist}
\alias{plotR_hist}
\title{Shows the histogram of temperatures with set of nests and the R function superimpose}
\usage{
plotR_hist(x, ..., ylimH = NULL, ylabH = "Frequency of temperatures")
}
\arguments{
\item{x}{Result data}

\item{...}{Parameters used by hist or plotR functions}

\item{ylimH}{Scale of histogram using ylimH=c(min, max)}

\item{ylabH}{Label for histogram scale}
}
\value{
Nothing
}
\description{
Shows the histogram of temperatures with set of nests and the R function superimpose
plotR_hist(data)
}
\details{
plotR_hist shows the histogram of temperatures with set of nests and the R function superimpose
}
\examples{
\dontrun{
library(embryogrowth)
data(resultNest_4p)
plotR_hist(resultNest_4p)
}
}
\author{
Marc Girondot \email{marc.girondot@u-psud.fr}
}

