% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultNest_4p_SSM_Linear.R
\docType{data}
\encoding{UTF-8}
\name{resultNest_4p_SSM_Linear}
\alias{resultNest_4p_SSM_Linear}
\title{Fit using the nest database}
\format{
A list with fitted information about data(nest)
}
\usage{
resultNest_4p_SSM_Linear
}
\description{
Fit using the nest database
}
\details{
Result of the fit using the nest database with linear progression
}
\examples{
\dontrun{
library(embryogrowth)
data(nest)
formated <- FormatNests(nest)
pfixed <- NULL
M0 = 0
############################################################################
# 4 parameters SSM
############################################################################
x <- c('DHA' = 64.868697530424186, 'DHH' = 673.18292743646771, 
       'T12H' = 400.90952554047749, 'Rho25' = 82.217237723502123)
resultNest_4p_SSM_Linear <- searchR(parameters=x, fixed.parameters=pfixed, 
	temperatures=formated, integral=integral.linear, M0=M0, 
	hatchling.metric=c(Mean=39.33, SD=1.92)/39.33)
}
}
\author{
Marc Girondot \email{marc.girondot@universite-paris-saclay.fr}
}
\keyword{datasets}
