% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{resultNest_4p_transition}
\alias{resultNest_4p_transition}
\title{Result of the fit using the nest database using transition}
\format{A list with fitted information about data(nest)}
\usage{
resultNest_4p_transition
}
\description{
Fit using the nest database using transition
}
\details{
Result of the fit using the nest database using transition
}
\examples{
\dontrun{
library(embryogrowth)
data(nest)
formated <- FormatNests(nest)
data(resultNest_4p_transition)
}
}
\author{
Marc Girondot \email{marc.girondot@u-psud.fr}
}
\references{
Girondot, M. & Kaska, Y. In press. A model to predict the thermal
         reaction norm for the embryo growth rate from field data. Journal of
         Thermal Biology. In press
}
\keyword{datasets}

