% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultNest_mcmc_4p_SSM4p.R
\docType{data}
\encoding{UTF-8}
\name{resultNest_mcmc_4p_SSM4p}
\alias{resultNest_mcmc_4p_SSM4p}
\title{Result of the mcmc using the nest database}
\format{
A list of class mcmcComposite with mcmc result for data(nest) with 4 parameters and Gompertz model of growth
}
\usage{
resultNest_mcmc_4p_SSM4p
}
\description{
Fit using the nest database
}
\details{
Result of the mcmc using the nest database
}
\examples{
\dontrun{
library(embryogrowth)
data(nest)
formated <- FormatNests(nest)
# The initial parameters value can be:
# "T12H", "DHA",  "DHH", "Rho25"
# Or
# "T12L", "DT", "DHA",  "DHH", "DHL", "Rho25"
x <- structure(c(118.431040984352, 498.205702157603, 306.056280989839, 
118.189669472381), .Names = c("DHA", "DHH", "T12H", "Rho25"))
# pfixed <- c(K=82.33) or rK=82.33/39.33
pfixed <- c(rK=2.093313)
resultNest_4p_SSM4p <- searchR(parameters=x, fixed.parameters=pfixed, 
	temperatures=formated, derivate=dydt.Gompertz, M0=1.7, 
	test=c(Mean=39.33, SD=1.92))
data(resultNest_4p_SSM4p)
pMCMC <- TRN_MHmcmc_p(resultNest_4p_SSM4p, accept=TRUE)
# Take care, it can be very long, sometimes several days
resultNest_mcmc_4p_SSM4p <- GRTRN_MHmcmc(result=resultNest_4p_SSM4p, 
 adaptive = TRUE, 
	parametersMCMC=pMCMC, n.iter=10000, n.chains = 1, n.adapt = 0,  
	thin=1, trace=TRUE)
data(resultNest_mcmc_4p_SSM4p)
1-rejectionRate(as.mcmc(resultNest_mcmc_4p_SSM4p))
as.parameters(resultNest_mcmc_4p_SSM4p)
layout(mat=matrix(1:4, nrow = 2))
plot(resultNest_mcmc_4p_SSM4p, parameters = "all", scale.prior = TRUE, las = 1)
layout(mat=1)
plotR(resultNest_4p_SSM4p, resultmcmc=resultNest_mcmc_4p_SSM4p, ylim=c(0,4), 
         main="Schoolfield, Sharpe & Magnuson 4-parameters", show.density=TRUE)
}
}
\references{
Girondot, M., & Kaska, Y. (2014). A model to predict 
            the thermal reaction norm for the embryo growth rate 
            from field data. Journal of Thermal Biology, 45, 96-102. 
            doi: 10.1016/j.jtherbio.2014.08.005
}
\author{
Marc Girondot \email{marc.girondot@u-psud.fr}
}
\keyword{datasets}
