% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROSIE.version.R
\name{ROSIE.version}
\alias{ROSIE.version}
\title{Version of database of TSD information for reptiles}
\usage{
ROSIE.version()
}
\value{
The date of the latest updated version
}
\description{
Return the date of the most recent update of the ROSIE database.
}
\details{
Database of information for incubation of turtles
}
\examples{
\dontrun{
library(embryogrowth)
ROSIE.version()
}
}
\seealso{
Other Functions for temperature-dependent sex determination: 
\code{\link{DatabaseTSD}},
\code{\link{DatabaseTSD.version}()},
\code{\link{P_TRT}()},
\code{\link{ROSIE}},
\code{\link{TSP.list}},
\code{\link{plot.tsd}()},
\code{\link{predict.tsd}()},
\code{\link{stages}},
\code{\link{tsd}()},
\code{\link{tsd_MHmcmc}()},
\code{\link{tsd_MHmcmc_p}()}
}
\author{
Marc Girondot \email{marc.girondot@universite-paris-saclay.fr}
}
\concept{Functions for temperature-dependent sex determination}
