% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STSRE_NestingArea.R
\docType{data}
\name{STSRE_NestingArea}
\alias{STSRE_NestingArea}
\title{Database of RMU for marine turtles}
\format{A dataframe with raw data.}
\usage{
STSRE_NestingArea
}
\description{
Database of RMU for marine turtles
}
\details{
Database of RMU for marine turtles
}
\examples{
\dontrun{
library(embryogrowth)
data(STSRE_NestingArea)
}
}
\author{
Marc Girondot \email{marc.girondot@u-psud.fr}

Maria Sousa Martins \email{maria.esmartins@gmail.com}
}
\references{
Wallace, B. P., DiMatteo, A. D., Hurley, B. J., Finkbeiner, 
            E. M., Bolten, A. B., Chaloupka, M. Y., Hutchinson, B. J., 
            Abreu-Grobois, F. A., Amorocho, D., Bjorndal, K. A., Bourjea, J., 
            Bowen, B. W., Duenas, R. B., Casale, P., Choudhury, B. C., Costa, A., 
            Dutton, P. H., Fallabrino, A., Girard, A., Girondot, M., 
            Godfrey, M. H., Hamann, M., Lopez-Mendilaharsu, M., Marcovaldi, 
            M. A., Mortimer, J. A., Musick, J. A., Nel, R., Seminoff, J. A., 
            Troeng, S., Witherington, B. & Mast, R. B. 2010. Regional management 
            units for marine turtles: a novel framework for prioritizing conservation 
            and research across multiple scales. Plos One, 5, e15465.
}
\keyword{datasets}

