% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/switch.transition.R
\name{switch.transition}
\alias{switch.transition}
\title{Add a transition parameter on a set of parameters or remove it}
\usage{
switch.transition(parameters = stop("A set of parameters must be supplied"))
}
\arguments{
\item{parameters}{A vector with parameters}
}
\value{
A vector with parameters
}
\description{
Add a transition parameter on a set of parameters or remove it
}
\details{
switch.transition Add a transition parameter on a set of parameters or remove it
}
\examples{
\dontrun{
data(resultNest_6p)
# Get a set of parameters without transition
x1 <- resultNest_6p$par
# Generate a set of parameters with transition
x2 <- switch.transition(x1)
# Generate a set of parameters without transition
x3 <- switch.transition(x3)
}
}
\author{
Marc Girondot
}

