% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.parallel.R
\name{test.parallel}
\alias{test.parallel}
\title{Estimate the likelihood of a set of parameters for nest incubation data with or without parallel computing option}
\usage{
test.parallel(result = stop("A ResultNest object must be provided"))
}
\arguments{
\item{result}{A object obtained after searchR or likelihoodR}
}
\value{
The gain or loss of computing time using parallel version
}
\description{
Estimate the likelihood of a set of parameters for nest incubation data with or without parallel computing option.
It uses the user time from the print result of system.time() function.
}
\details{
test.parallel estimates the likelihood of a set of parameters for nest incubation data with or without parallel computing option
}
\examples{
\dontrun{
library(embryogrowth)
data(resultNest_4p_SSM4p)
test.parallel(resultNest_4p_SSM4p)
}
}
\author{
Marc Girondot
}
