% Generated by roxygen2 (4.0.1): do not edit by hand
\name{TestParallel}
\alias{TestParallel}
\title{Estimate the likelihood of a set of parameters for nest incubation data with or without parallel computing option}
\usage{
TestParallel(result = stop("A ResultNest object must be provided"))
}
\arguments{
\item{result}{A object obtained after searchR or likelihoodR}
}
\value{
Nothing
}
\description{
Estimate the likelihood of a set of parameters for nest incubation data with or without parallel computing option
}
\details{
TestParallel estimates the likelihood of a set of parameters for nest incubation data with or without parallel computing option
}
\examples{
\dontrun{
library(embryogrowth)
data(resultNest_4p)
TestParallel(resultNest_4p)
}
}
\author{
Marc Girondot
}

