% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{nest}
\alias{nest}
\title{Timeseries of temperatures for nests}
\format{A dataframe with raw data.}
\usage{
nest
}
\description{
Timeseries of temperatures for nests
}
\details{
Timeseries of temperatures for nests
}
\examples{
\dontrun{
library(embryogrowth)
data(nest)
}
}
\author{
Marc Girondot \email{marc.girondot@u-psud.fr}
}
\references{
Girondot, M. & Kaska, Y. Submitted. A model to predict
            temperature dependency on embryo growth rate and incubation
            duration from field data.
}
\keyword{datasets}

