% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.tsd.R
\name{plot.tsd}
\alias{plot.tsd}
\title{Plot results of tsd() that best describe temperature-dependent sex determination}
\usage{
\method{plot}{tsd}(x, ..., males.freq = TRUE, las.x = 1, las.y = 1,
  lab.PT = "Pivotal temperature",
  lab.TRT = paste0("Transitional range of temperatures l=", l * 100, "\%"),
  col.TRT = "gray", col.TRT.CI = rgb(0.8, 0.8, 0.8, 0.5),
  col.PT.CI = rgb(0.8, 0.8, 0.8, 0.5), show.CI = TRUE)
}
\arguments{
\item{x}{A result file generated by tsd()}

\item{...}{Parameters for plot()}

\item{males.freq}{Should the graph uses males frequency [TRUE] or females [FALSE]}

\item{las.x}{las parameter for x axis}

\item{las.y}{las parameter for y axis}

\item{lab.PT}{Label to describe pivotal temperature}

\item{lab.TRT}{Label to describe transitional range of temperature}

\item{col.TRT}{The color of TRT}

\item{col.TRT.CI}{The color of CI of TRT based on range.CI}

\item{col.PT.CI}{The color of CI of PT based on range.CI}

\item{show.CI}{Do the CI for the curve should be shown}
}
\value{
Nothing
}
\description{
Plot the estimates that best describe temperature-dependent sex determination.
}
\details{
plot.tsd plot result of tsd() that best describe temperature-dependent sex determination
}
\examples{
\dontrun{
CC_AtlanticSW <- subset(STSRE_TSD, RMU=="Atlantic, SW" &
                          Species=="Caretta caretta" & Sexed!=0)
tsdL <- with (CC_AtlanticSW, tsd(males=Males, females=Females,
                                 temperatures=Incubation.temperature-Correction.factor,
                                 equation="logistic"))
plot(tsdL)
}
}
\author{
Marc Girondot
}
\references{
Girondot, M. 1999. Statistical description of temperature-dependent sex determination using maximum likelihood. Evolutionary Ecology Research, 1, 479-486.

Godfrey, M.H., Delmas, V., Girondot, M., 2003. Assessment of patterns of temperature-dependent sex determination using maximum likelihood model selection. Ecoscience 10, 265-272.

Hulin, V., Delmas, V., Girondot, M., Godfrey, M.H., Guillon, J.-M., 2009. Temperature-dependent sex determination and global change: are some species at greater risk? Oecologia 160, 493-506.

Girondot M., Submited. On the concept of embryological thermosensitive period for sex determination in reptiles.
}

