% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultNest_4p_transition.R
\docType{data}
\encoding{UTF-8}
\name{resultNest_4p_transition}
\alias{resultNest_4p_transition}
\title{Result of the fit using the nest database using transition}
\format{A list with fitted information about data(nest)}
\usage{
resultNest_4p_transition
}
\description{
Fit using the nest database using transition
}
\details{
Result of the fit using the nest database using transition
}
\examples{
\dontrun{
library(embryogrowth)
data(nest)
formated <- FormatNests(nest)
data(resultNest_4p_transition)
}
}
\references{
Girondot, M., & Kaska, Y. (2014). A model to predict 
            the thermal reaction norm for the embryo growth rate 
            from field data. Journal of Thermal Biology, 45, 96-102. 
            doi: 10.1016/j.jtherbio.2014.08.005
}
\author{
Marc Girondot \email{marc.girondot@u-psud.fr}
}
\keyword{datasets}
