% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultNest_newp.R
\docType{data}
\encoding{UTF-8}
\name{resultNest_newp}
\alias{resultNest_newp}
\title{Fit using the nest database with anchored parameters}
\format{
A list with fitted information from data(nest) with anchored parameters
}
\usage{
resultNest_newp
}
\description{
Fit using the nest database with anchored parameters
}
\details{
Result of the fit using the nest database with anchored parameters
}
\examples{
\dontrun{
library(embryogrowth)
data(nest)
formated <- FormatNests(nest)
newp <- GenerateAnchor(nests=formated, number.anchors=7)
pfixed <- c(rK=2.093313)
resultNest_newp <- searchR(parameters=newp, fixed.parameters=pfixed, 
  temperatures=formated, integral=integral.Gompertz, M0=1.7, 
	hatchling.metric=c(Mean=39.33, SD=1.92))
data(resultNest_newp)
plotR(resultNest_newp)
}
}
\references{
Girondot, M., & Kaska, Y. (2014). A model to predict
the thermal reaction norm for the embryo growth rate
from field data. Journal of Thermal Biology, 45, 96-102.
doi: 10.1016/j.jtherbio.2014.08.005
}
\author{
Marc Girondot \email{marc.girondot@universite-paris-saclay.fr}
}
\keyword{datasets}
