% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Nests.R
\name{summary.Nests}
\alias{summary.Nests}
\title{Summarize the information from a Nests object.}
\usage{
\method{summary}{Nests}(object, ...)
}
\arguments{
\item{object}{A object obtained after FormatNests()}

\item{...}{Not used}
}
\value{
None
}
\description{
Summarize the information from a Nests object:\cr
- Name of the nests, total incubation length and average temperature
}
\details{
summary.Nests Summarize the information from a Nests object
}
\examples{
\dontrun{
library(embryogrowth)
data(nest)
formated <- FormatNests(nest, previous=NULL)
summary(formated)
}
}
\author{
Marc Girondot
}
