% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TestParallel.R
\name{TestParallel}
\alias{TestParallel}
\title{Estimate the likelihood of a set of parameters for nest incubation data with or without parallel computing option}
\usage{
TestParallel(result = stop("A ResultNest object must be provided"))
}
\arguments{
\item{result}{A object obtained after searchR or likelihoodR}
}
\value{
The gain or loss of computing time using parallel version
}
\description{
Estimate the likelihood of a set of parameters for nest incubation data with or without parallel computing option.
It uses the user time from the print result of system.time() function.
}
\details{
TestParallel estimates the likelihood of a set of parameters for nest incubation data with or without parallel computing option
}
\examples{
\dontrun{
library(embryogrowth)
data(resultNest_4p)
TestParallel(resultNest_4p)
}
}
\author{
Marc Girondot
}

