% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.tsd.R
\name{web.tsd}
\alias{web.tsd}
\title{Run a shiny application for basic functions of tsd function}
\usage{
web.tsd()
}
\value{
Nothing
}
\description{
Run a shiny application for basic functions of tsd function.
}
\details{
web.tsd runs a shiny application for basic functions of tsd function
}
\examples{
\dontrun{
library(embryogrowth)
web.tsd()
}
}
\author{
Marc Girondot
}
